Show response_timestamp in Invoice Rejection Detail window

+++ backend/vertical-healthcare_ch/forum_datenaustausch_ch.invoice_450_response/src/main/java/de/metas/vertical/healthcare_ch/forum_datenaustausch_ch/invoice_450/response/Invoice450ToCrossVersionModelTool.java
@@ -76,7 +76,7 @@ public class Invoice450ToCrossVersionModelTool
 		final XmlPayloadBuilder payloadBuilder = XmlPayload.builder();
 
 		payloadBuilder.invoice(createXmlInvoice(payload.getInvoice()));
-		payloadBuilder.responseTimestamp(payload.getResponseTimestamp());
+		payloadBuilder.responseTimestamp(BigInteger.valueOf(payload.getResponseTimestamp()));
 		payloadBuilder.body(createXmlBody(payload.getBody()));
 
 		return payloadBuilder.build();
