+++ backend/vertical-healthcare_ch/forum_datenaustausch_ch.invoice_base/src/main/java/de/metas/vertical/healthcare_ch/forum_datenaustausch_ch/base/export/invoice/InvoiceImportClientImpl.java
@@ -102,7 +102,7 @@ public class InvoiceImportClientImpl implements InvoiceImportClient
 
 		final Instant invoiceCreatedTimestamp = Instant.ofEpochSecond(invoice.getRequestTimestamp().longValue());
 
-		final Instant invoiceResponseTimestamp = Instant.ofEpochSecond(payload.getResponseTimestamp());
+		final Instant invoiceResponseTimestamp = Instant.ofEpochSecond(payload.getResponseTimestamp().longValue());
 
 		final ImportedInvoiceResponseBuilder builder = ImportedInvoiceResponse
 				.builder()
